Move implementation functions to api-helper
